pkg/ring: Add state to tokens
This commit adds state to each token alongside the overall ingester
state. The ingester's state is kept for backwards compatibility and
to making it easier to see what the ingester is doing (i.e., in health
checks or on the ring status web page). Ingesters also store a boolean
indicating whether they have stateful tokens.

The Get method on the ring has been modified to use the token state
rather than ingester state. When reading the ring, if an ingester
doesn't have stateful tokens enabled, token states are set to
match the ingester's state.

Currently, the state of the tokens are forced to have the same state as
the ingester. Further changes will separate out the two concepts.

As old gRPC clients still need to access normalised tokens, the
protobuf could not be updated to change the type of the tokens field.
Rather, an extra "inactiveTokens" field has been added to IngesterDesc.
This field is a map of all non-ACTIVE tokens to the state they are
currently in.

Signed-off-by: Robert Fratto <robert.fratto@grafana.com>

pkg/ring: incremental transfer management

This commit introduces managing incremental transfers between ingesters
when a lifecycler joins a ring and when it leaves the ring. The
implementation of the IncrementalTransferer interface will be done in a
future commit.

The LifecyclerConfig has been updated with JoinIncrementalTransfer and
LeaveIncrementalTransfer, available as join_incremental_transfer and
leave_incremental_transfer using the YAML config, and
join-incremental-transfer and leave-incremental-transfer using command
line flags.

When JoinIncrementalTransfer is used, the lifecycler will join the ring
immediately. Tokens will be inserted into the ring one by one, first
into the JOINING state and then the ACTIVE state after requesting chunks
in token ranges they should have data for from neighboring ingesters
in the ring.

When LeaveIncrementalTransfer is used, the lifecycler will incrementally
move tokens in LEAVING state after sending ranges to neighboring
ingesters that should now have data. Enabling LeaveIncrementalTransfer
will disable the handoff process, and flushing non-transferred data always
happens at the end.

Signed-off-by: Robert Fratto <robert.fratto@grafana.com>

pkg/distributor, pkg/ingester: push shard token to ingester

This commit modifies the ingesters to be aware of the shard token used by
the distributors to send traffic to ingesters. This is a requirement for
incremental transfers, where the shard token is used to determine which
memory series need to be moved.

This assumes that all distributors are using the same sharding mechanism
and always use the same token for a specific series. If the memory
series is appended to with a different token from the one it was created
with, a warning will be logged and the new token will be used.

Signed-off-by: Robert Fratto <robert.fratto@grafana.com>

pkg/ingester: implement IncrementalTransferer interface

This commit implements the IncrementalTransferer interface used by
lifecyclers to move memory series around the ring as ingesters join and
leave.

Signed-off-by: Robert Fratto <robert.fratto@grafana.com>

pkg/ring: add TokenChecker

This commit introduces a TokenChecker component which runs in the
background to support reporting metrics on unexpected tokens pushed to
ingesters. It supports checking on an interval, checking when a new
stream is pushed, checking when an existing stream is appended to,
and checking when a stream is transferred.

Signed-off-by: Robert Fratto <robert.fratto@grafana.com>

Review feedback and minor refactoring

- add method for token range checking key
- move incremental transfer code to new files
- simplify ring.Desc.IsRange implementation
- fix remaining spillover on incremental transfer join
- improve performance of retrieving token predecessors

Signed-off-by: Robert Fratto <robert.fratto@grafana.com>

Address review feedback, refactor some code

Signed-off-by: Robert Fratto <robert.fratto@grafana.com>

Address review feedback

Signed-off-by: Robert Fratto <robertfratto@gmail.com>

pkg/ring: make Desc.search less confusing

Desc.search has been renamed to Desc.findToken. This helps it seem
less similar to Ring.search which searches for an owning token of a
key rather than a specific token.

Signed-off-by: Robert Fratto <robert.fratto@grafana.com>

Address some review feedback

Signed-off-by: Robert Fratto <robert.fratto@grafana.com>

pkg/ring: address review feedback

Signed-off-by: Robert Fratto <robert.fratto@grafana.com>

pkg/ring: first pass at reverting stateful tokens

This commit reverts the addition of stateful tokens. Everything works in
my local environment, and all the tests pass, but there may be more
opporuntities to remove code that stateful tokens added. Work for that
will be done in a separate commit.

Signed-off-by: Robert Fratto <robert.fratto@grafana.com>

pkg/ring: refactor, remove dead code

Signed-off-by: Robert Fratto <robert.fratto@grafana.com>

Address more review feedback

Signed-off-by: Robert Fratto <robert.fratto@grafana.com>

pkg/ring: address review feedback

Signed-off-by: Robert Fratto <robert.fratto@grafana.com>

pkg/ring: address review feedback

migrateRing has been renamed to NewTokenNavigator. References to
migrateRing that updated the token list of ring.Desc have been changed
to leave the ring unmodified and store a separate reference to the
TokenNavigator.

Signed-off-by: Robert Fratto <robertfratto@gmail.com>

pkg/ring: address extra minor feedback

Signed-off-by: Robert Fratto <robert.fratto@grafana.com>

pkg/ring: remove unneeded Desc.Clone method

Signed-off-by: Robert Fratto <robert.fratto@grafana.com>

pkg/ring: remove unneeded flags

Signed-off-by: Robert Fratto <robert.fratto@grafana.com>

pkg/ring: fix issues with incremental join/leave

- incremental join was assuming tokens were stored denormalized in the
  ring
- incremental join/leave weren't updating their copy of the ring, so if
  the join/leave process took more than the heartbeat timeout, the
  ingester would think all other ingesters were unhealthy and not
  heartbeating.

Signed-off-by: Robert Fratto <robert.fratto@grafana.com>

pkg/ingester: use 503 Service Unavailable during blocked range

The 503 error code makes more sense to return to clients rather than
temporarily returning a 4xx and forcing it back to a 5xx before
returning to the client.

Signed-off-by: Robert Fratto <robert.fratto@grafana.com>