pkg/ring: Add state to tokens
This commit adds state to each token alongside the overall ingester
state. The ingester's state is kept for backwards compatibility and
to making it easier to see what the ingester is doing (i.e., in health
checks or on the ring status web page). Ingesters also store a boolean
indicating whether they have stateful tokens.

The Get method on the ring has been modified to use the token state
rather than ingester state. When reading the ring, if an ingester
doesn't have stateful tokens enabled, token states are set to
match the ingester's state.

Currently, the state of the tokens are forced to have the same state as
the ingester. Further changes will separate out the two concepts.

As old gRPC clients still need to access normalised tokens, the
protobuf could not be updated to change the type of the tokens field.
Rather, an extra "inactiveTokens" field has been added to IngesterDesc.
This field is a map of all non-ACTIVE tokens to the state they are
currently in.

Signed-off-by: Robert Fratto <robert.fratto@grafana.com>