pkg/distributor, pkg/ingester: push shard token to ingester
This commit modifies the ingesters to be aware of the shard token used by
the distributors to send traffic to ingesters. This is a requirement for
incremental transfers, where the shard token is used to determine which
memory series need to be moved.

This assumes that all distributors are using the same sharding mechanism
and always use the same token for a specific series. If the memory
series is appended to with a different token from the one it was created
with, a warning will be logged and the new token will be used.

Signed-off-by: Robert Fratto <robert.fratto@grafana.com>