pkg/ring, pkg/ingester: add method for token range checking key
A TokenRange.Contains method has been added to simplify the
implementation of passing TokenRanges around. Previously, if a range
wrapped the ring (i.e., the From value was hire than the To value),
code creating ranges would unwrap that into two ranges: one from
From to math.MaxUint32 and another from 0 to To.

The Contains method catches this edge case and handles it, avoiding
the need of unwrapping a range.

Signed-off-by: Robert Fratto <robert.fratto@grafana.com>