pkg/ring: fix remaining spillover on incremental transfer join
findTransferWorkload was incorrectly discovering the target token when
tokens to the right of the joining/leaving token are already in the
replica set. Because the target was being found from the joining/leaving
token, Successor was unaware of ingesters starting from endRange.

The fix for this simplifies the logic: the target token is simplify
ReplicationFactor number of ingesters clockwise from the startRange
(not endRange) token.

Signed-off-by: Robert Fratto <robert.fratto@grafana.com>