pkg/ring: improve performance of retrieving token predecessors
The Predecessors function has been re-written with a slightly less naive
solution: starting at the token we want to get the predecessors for,
move in counter-clockwise order through the ring until we have seen N+1
unique ingesters. For each token we come across, if their Nth successor
is the starting token, we add them as a predecessor.

The new approach is similar to the original approach but cuts down the
search size down dramatically.

Signed-off-by: Robert Fratto <robert.fratto@grafana.com>